+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support for `--ignored-suffix` argument. See #1892 (@bojan88)
 - `$BAT_CONFIG_DIR` is now a recognized environment variable. It has precedence over `$XDG_CONFIG_HOME`, see #1727 (@billrisher)
 - Support for `x:+delta` syntax in line ranges (e.g. `20:+10`). See  #1810 (@bojan88)
+- Add new `--acknowledgements` argument that gives credit to theme and syntax definition authors. See #1971 (@Enselic)
 
 ## Bugfixes